Cleaning up

+++ plotting.py
@@ -37,32 +37,42 @@ def save_image(image, plot_title, output_dir):
     plt.close()
 
 
-def plot_loss_log(loss_log_path, result_dir):
+def plot_loss_log(loss_log_path, validation_loss_log_path, result_dir):
     """
     Reads the loss log file and plots the losses as a semilogy plot.
 
     :param loss_log_path: Path to the loss log file. It needs to have columns "epoch", "batch", and "average_loss".
+    :param validation_loss_log_path: Path to validation loss log file. It needs to have columns "epoch", "batch", and
+    "average_validation_loss"
     :param result_dir: Directory to save the plot.
     """
     if os.path.exists(loss_log_path):
-        # Read the loss log file
-        loss_data = pd.read_csv(loss_log_path)
-        epochs = loss_data['epoch']
-        batches = loss_data['batch']
-        losses = loss_data['average_loss']
+        if os.path.exists(validation_loss_log_path):
+            loss_data = pd.read_csv(loss_log_path)
+            epochs = loss_data['epoch']
+            losses = loss_data['average_loss']
 
-        # Plot the losses
-        plt.figure()
-        plt.semilogy(epochs + batches / batches.max(), losses, label='Loss')  # Combine epoch and batch for x-axis
-        plt.xlabel('Epoch + Batch (normalized)')
-        plt.ylabel('Loss (log scale)')
-        plt.title('Training Loss')
-        plt.legend()
-        plt.grid(True)
-        plt.savefig(os.path.join(result_dir, 'losses_plot.png'))
-        plt.close()
+            validation_loss_data = pd.read_csv(validation_loss_log_path)
+            # Use correct columns for validation loss
+            validation_epochs = validation_loss_data['epoch']
+            validation_losses = validation_loss_data['average_validation_loss']
+
+            # Plot the losses
+            plt.figure()
+            # Combine epoch and batch for x-axis, we add the +1 since the batches are counted with their literal index
+            plt.semilogy(epochs, losses, label='Training Loss')
+            plt.semilogy(validation_epochs, validation_losses, label='Validation Loss')
+            plt.xlabel('epochs')
+            plt.ylabel('loss (log scale)')
+            plt.title('Training and Validation Loss')
+            plt.legend()
+            plt.grid(True)
+            plt.savefig(os.path.join(result_dir, 'losses_plot.png'))
+            plt.close()
+        else:
+            raise Exception(f"Validation loss log file not found at {validation_loss_log_path}")
     else:
-        print(f"Loss log file not found at {loss_log_path}")
+        raise Exception(f"Loss log file not found at {loss_log_path}")
 
 
 def compare_images(image1, image2, file_name, output_location, title1, title2):
@@ -97,11 +107,11 @@ def compare_images(image1, image2, file_name, output_location, title1, title2):
 
 def draw_output_boxes(ax, particle_locations, image_height, image_width, box_color):
     for coords in particle_locations:
-        box_width = coords[2]
-        box_height = coords[3]
+        box_width = coords[2].item()
+        box_height = coords[3].item()
 
-        x = coords[0] - box_width / 2
-        y = coords[1] + box_height / 2  # Plus here because patches.Rectangle draws boxes assuming the origin is the bottom left
+        x = coords[0] - box_width
+        y = coords[1] + box_height  # Plus here because patches.Rectangle draws boxes assuming the origin is the bottom left
 
         # We do this weird image_height - y here because patches draws assuming the origin is in the bottom left corner
         box = patches.Rectangle((x, image_height - y), box_width, box_height, linewidth=1, edgecolor=box_color, facecolor='none')
@@ -110,8 +120,8 @@ def draw_output_boxes(ax, particle_locations, image_height, image_width, box_col
 
 def draw_bounding_boxes(ax, particle_locations, image_height, image_width, box_width, box_height, box_color):
     for coords in particle_locations:
-        x = coords[0] - box_width / 2
-        y = coords[1] + box_height / 2  # Box height here because box is drawn based on the top left corner
+        x = coords[0] - box_width
+        y = coords[1] + box_height  # Box height here because box is drawn based on the top left corner
 
         # We do this weird image_height - y here because patches draws assuming the origin is in the bottom left corner
         box = patches.Rectangle((x, image_height - y), box_width, box_height, linewidth=1, edgecolor=box_color, facecolor='none')
@@ -231,14 +241,14 @@ def compare_predictions_with_ground_truth(image_tensor, ground_truth, prediction
     plt.savefig(os.path.join(result_dir, file_name))
 
 
-def compare_heatmaps_with_ground_truth(micrograph, particle_locations, heatmaps, heatmaps_title, result_folder_name,
-                                       result_dir):
-    create_folder_if_missing(os.path.join(result_dir, result_folder_name))
+def compare_heatmaps_with_ground_truth(micrograph, particle_locations, heatmaps, heatmaps_title,
+                                       result_dir, result_file_name):
+    create_folder_if_missing(os.path.join(result_dir))
 
     fig, (ax1, ax2) = plt.subplots(1, 2)
     ax1.set_title('Ground Truth')
 
-    draw_image_with_objects_on_ax(ax1, micrograph, particle_locations, "output_box", {},
+    draw_image_with_objects_on_ax(ax1, micrograph, particle_locations, "circle", {"circle_radius": 6},
                                   "r")
 
     for i in range(len(heatmaps)):
@@ -247,7 +257,7 @@ def compare_heatmaps_with_ground_truth(micrograph, particle_locations, heatmaps,
 
         cbar2 = fig.colorbar(im2, ax=ax2, orientation='vertical', shrink=0.5)
 
-        plt.savefig(os.path.join(result_dir, result_folder_name, f"heatmap_{i}_compared_to_ground_truth.png"))
+        plt.savefig(os.path.join(result_dir, result_file_name))
 
         cbar2.remove()
         ax2.cla()  # Clear before doing the next comparison
